+++ src/test/java/com/pivovarit/collectors/ParallelismTest.java
@@ -9,9 +9,13 @@ import org.junit.runner.RunWith;
 
 import java.time.Duration;
 import java.util.List;
+import java.util.concurrent.TimeUnit;
+import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
 import static com.pivovarit.collectors.ParallelCollectors.parallel;
+import static com.pivovarit.collectors.ParallelCollectors.parallelToOrderedStream;
+import static com.pivovarit.collectors.ParallelCollectors.parallelToStream;
 import static com.pivovarit.collectors.infrastructure.TestUtils.TRIALS;
 import static com.pivovarit.collectors.infrastructure.TestUtils.returnWithDelay;
 import static com.pivovarit.collectors.infrastructure.TestUtils.returnWithDelayGaussian;
@@ -27,6 +31,33 @@ import static org.awaitility.Awaitility.await;
 public class ParallelismTest extends ExecutorAwareTest {
 
     @Property(trials = TRIALS)
+    public void shouldCollectWithThrottledParallelism(@InRange(minInt = 20, maxInt = 100) int unitsOfWork, @InRange(minInt = 1, maxInt = 20) int parallelism) {
+        // given
+        TestUtils.CountingExecutor executor = new TestUtils.CountingExecutor();
+
+        Stream.generate(() -> 42)
+          .limit(unitsOfWork)
+          .collect(parallelToStream(i -> returnWithDelay(42L, ofMillis(Integer.MAX_VALUE)), executor, parallelism));
+
+        await()
+          .pollDelay(100, TimeUnit.MILLISECONDS)
+          .until(() -> executor.count() == parallelism);
+    }
+
+    @Property(trials = TRIALS)
+    public void shouldMaintainOrder(@InRange(minInt = 2, maxInt = 20) int unitsOfWork, @InRange(minInt = 2, maxInt = 40) int parallelism) {
+        // given
+        executor = threadPoolExecutor(unitsOfWork);
+        List<Integer> result = Stream.iterate(0, i -> i + 1).limit(20)
+          .collect(parallelToOrderedStream(i -> returnWithDelayGaussian(i, Duration.ofMillis(10)), executor, parallelism))
+          .collect(Collectors.toList());
+
+        assertThat(result).isSorted();
+
+        executor.shutdownNow();
+    }
+
+    @Property(trials = TRIALS)
     public void shouldCollectToListWithThrottledParallelism(@InRange(minInt = 20, maxInt = 100) int unitsOfWork, @InRange(minInt = 1, maxInt = 20) int parallelism) {
         // given
         TestUtils.CountingExecutor executor = new TestUtils.CountingExecutor();
@@ -40,7 +71,7 @@ public class ParallelismTest extends ExecutorAwareTest {
     }
 
     @Property(trials = TRIALS)
-    public void shouldMaintainOrder(@InRange(minInt = 2, maxInt = 20) int unitsOfWork, @InRange(minInt = 2, maxInt = 40) int parallelism) {
+    public void shouldCollectToListAndMaintainOrder(@InRange(minInt = 2, maxInt = 20) int unitsOfWork, @InRange(minInt = 2, maxInt = 40) int parallelism) {
         // given
         executor = threadPoolExecutor(unitsOfWork);
         List<Integer> result = Stream.iterate(0, i -> i + 1).limit(20)
